+++ mypyc/irbuild/for_helpers.py
